component renaming and removal of obsolet components
